requirements.txt
django-sites
django-allauth
